Added navbar and major render fix
While rendering dashboard page, it was a bug when logging out and trying to access dashboard again, the app was rendering the dashboard(tho it shouldnt). Fixed this by changing rendering condition in dashboard.js and login.js.
Also added minor improvement(before the app was rendering the login page and redirecting to dashboard if logged) now it only renders the dashboard page if logged and login page if not logged